requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator
easyocr
numpy
Pillow
opencv-python
textblob
transformers
